face-recognition-service/requirements.txt
# Windows-compatible requirements for face recognition service
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
Pillow==10.0.0

# Use pre-compiled wheels for Windows
dlib==19.24.2
face-recognition==1.3.0

# Additional utilities
python-dotenv==1.0.0